irony_backend/requirements.txt
fastapi[all]
uvicorn
motor
requests
joblib
apscheduler
passlib
bcrypt
pyjwt
pymongo>=4.0.0
redis